# fix: msecs option in v7