@broofa 2024-06-18
@robinpokorny Sorry for the delay on this.  I'm kind of hoping @pmccarren will chime in here.

If this is very-urgent and pmccaren deems it appropriate we can push this out.  But I'm not thrilled about adding yet-more internal state variables without being a bit more deliberate in how we handle them.  I'm thinking it might be worth refactoring this code to formalize the internal state as a data structure that is used both internally or that can be optionally passed as an option.  And then break the v6() method into two parts: 
1. update internal state
2. generate uuid from internal state

That might be a breaking change though...?  So maybe we push this out, then actually-fix how we handle internal state separately and maybe push that out with the breaking-changes in the TS port work.